Small changes
After changing the default layer I wasn't seeing the line until
triggering a data update. To fix this I added calls to `setupMap` to
remove the line if it was there, and `addDataToMap` to re-add the line
after the layer was changed.

I also made the form size bit bigger since it was cutting off the
longest layer name.