Add basic GeoJSON overlay
This is a first test to add a textarea in which a user can copy-paste GeoJSOn code.
It is not finished, the style may be not great but more importantly the overlay
is not erased if the code is erased. Overlays are only added and never removed.

The next step is to add the import button functionnality.