Add list view for overlays
Now when the user hits the "Save" button, the content of the textarea is
saved as a new element of the overlay list and the textarea is cleared.
This list is displayed on the right side and for each element is
displayed: its name (for the moment its Leaflet id) and three buttons.
These three buttons are: download (to download the overlay as a file),
edit (to make the content of the overlay appear in the text area) and
delete (to delete the overlay from the list and from the map). These
buttons don't do anything for the moment.