Fix dependencies
Note that if the dependencies of `@babel/preset-env` are updated then
the plugin fails to build properly (Grafana error: "TypeError: can't
access property "events", this.panel is undefined").

Right now the `package-lock.json` is holding the versions of the deps in
place so that doesn't happen but this should be fixed properly in the
future.

Tested and working in Grafana 7.2.1 so closes #32 .